# api-docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "YukiBloom 网站的文章、图片、个人内容和站点页面 API。JSON 响应统一使用 code、message、data；成功 code 为 0，失败 code 与 HTTP 状态码一致。",
        "title": "YukiBloom 后端接口",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/posts": {
            "get": {
                "description": "需要先通过管理员登录接口登录，浏览器会自动携带会话 Cookie。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理文章"
                ],
                "summary": "查询后台文章列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量，默认 20，最大 100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文章分页结果，包含草稿",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/posts.Page"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "分页参数无效",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            },
            "post": {
                "description": "需要登录；写请求需在 X-CSRF-Token 请求头传入登录响应中的 csrfToken。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理文章"
                ],
                "summary": "创建文章草稿",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录接口返回的 csrfToken",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "文章内容",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/posts.PostInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建后的草稿",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/posts.Post"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求内容无效或封面图片不存在",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "403": {
                        "description": "CSRF 校验失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "409": {
                        "description": "语言和链接标识已被占用",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "保存失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/posts/{id}": {
            "get": {
                "description": "需要先通过管理员登录接口登录。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理文章"
                ],
                "summary": "查询后台文章详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章 UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文章详情，包含草稿正文和版本号",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/posts.Post"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "文章 ID 格式错误",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            },
            "patch": {
                "description": "需要登录和 CSRF 令牌。请求中的 version 必须与读取文章时一致，否则返回冲突。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理文章"
                ],
                "summary": "修改文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章 UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录接口返回的 csrfToken",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "文章新内容及当前 version",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/posts.PostInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新后的文章",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/posts.Post"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求内容、文章 ID 或封面图片无效",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "403": {
                        "description": "CSRF 校验失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "409": {
                        "description": "文章已被其他操作修改或链接标识冲突",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "保存失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/posts/{id}/publish": {
            "post": {
                "description": "需要登录和 CSRF 令牌。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理文章"
                ],
                "summary": "发布文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章 UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录接口返回的 csrfToken",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "已发布文章",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/posts.Post"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "文章 ID 格式错误",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "403": {
                        "description": "CSRF 校验失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "发布失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/posts/{id}/unpublish": {
            "post": {
                "description": "需要登录和 CSRF 令牌。撤回后文章不再出现在公开接口。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理文章"
                ],
                "summary": "撤回文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章 UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录接口返回的 csrfToken",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "已撤回文章",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/posts.Post"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "文章 ID 格式错误",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "401": {
                        "description": "尚未登录",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "403": {
                        "description": "CSRF 校验失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "撤回失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/posts/{slug}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "按链接标识查询公开文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章链接标识",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "语言代码，例如 zh-CN",
                        "name": "locale",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文章详情",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresponse.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/posts.Post"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "语言代码无效",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/apiresponse.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiresponse.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "posts.Page": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/posts.Post"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "posts.PostInput": {
            "type": "object",
            "properties": {
                "bodyMarkdown": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "coverMediaId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayDate": {
                    "type": "string"
                },
                "extra": {
                    "type": "object"
                },
                "locale": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    }
}